+++ src/core/pipeline/items/string/generation/cuid2.rs
@@ -1,22 +1,22 @@
 use async_trait::async_trait;
-use cuid::slug;
+use cuid2::create_id;
 use crate::core::pipeline::item::Item;
 use crate::core::teon::Value;
 use crate::core::result::Result;
 use crate::core::pipeline::ctx::Ctx;
 
 #[derive(Debug, Copy, Clone)]
-pub struct SlugItem {}
+pub struct CUID2Item {}
 
-impl SlugItem {
+impl CUID2Item {
     pub fn new() -> Self {
-        SlugItem {}
+        Self {}
     }
 }
 
 #[async_trait]
-impl Item for SlugItem {
+impl Item for CUID2Item {
     async fn call<'a>(&self, ctx: Ctx<'a>) -> Result<Ctx<'a>> {
-        Ok(ctx.with_value(Value::String(slug().unwrap())))
+        Ok(ctx.with_value(Value::String(create_id())))
     }
 }