+++ src/core/pipeline/items/string/generation/mod.rs
@@ -2,3 +2,4 @@ pub mod cuid;
 pub mod uuid;
 pub mod random_digits;
 pub mod slug;
+pub mod cuid2;